refactor: 장바구니 및 결제 관련 컴포넌트 구조화
- Cart 컴포넌트를 추가하여 장바구니 UI를 통합
- Checkout 및 CouponSelector 컴포넌트를 추가하여 결제 정보 및 쿠폰 선택 기능 분리
- App.tsx에서 장바구니 관련 JSX를 Cart 컴포넌트로 대체하여 코드 가독성 향상
- 불필요한 코드 삭제 및 구조 개선으로 유지보수성 향상
